+++ src/volumeobject.cpp
@@ -66,9 +66,9 @@ QStringList VolumeObject::channels() const
     return m_channels;
 }
 
-QList<qint64> VolumeObject::channelVolumes() const
+QList<qreal> VolumeObject::channelVolumes() const
 {
-    QList<qint64> ret;
+    QList<qreal> ret;
     ret.reserve(m_volume.channels);
     for (int i = 0; i < m_volume.channels; ++i) {
         ret << m_volume.values[i];
